Rename InputHandler
